% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(data)
}
\arguments{
\item{data}{object that was retrieved from Twarc2}
}
\value{
Removes unnecessary variables, removes accounts with too many or too less friends,
}
\description{
Clean Data
}
\examples{
quick_sum(twarc_data)
}
