% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{location_reformat}
\alias{location_reformat}
\title{Re-format Location Info}
\usage{
location_reformat(data)
}
\arguments{
\item{data}{object that was retrieved from Twarc2}
}
\value{
This approach is based on TIMME's location re-formatting. Using the public location info, it creates three new columns:"city","state"
}
\description{
Re-format Location Info
}
\examples{
quick_sum(twarc_data)
}
