% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot Timeline of the Tweets (Rtweet)}
\usage{
plot_frequency(data)
}
\arguments{
\item{data}{object that was retrieved from Twarc2}
}
\value{
USER RTWEET. This is now deprecated. Plot the frequency change of tweets over time. This is directly borrowed from rtweet's `ts_plot` function.
}
\description{
Plot Timeline of the Tweets (Rtweet)
}
\examples{
quick_sum(twarc_data)
}
