% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{plot_hourly}
\alias{plot_hourly}
\title{Plot Timeline of the Tweets (by Hour)}
\usage{
plot_hourly(data)
}
\arguments{
\item{data}{object that was retrieved from Twarc2}
}
\value{
Plot the frequency change of tweets over time.
}
\description{
Plot Timeline of the Tweets (by Hour)
}
\examples{
quick_sum(twarc_data)
}
