% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{quick_sum}
\alias{quick_sum}
\title{Quick Summary of Tweets}
\usage{
quick_sum(data)
}
\arguments{
\item{data}{Csv object that was retrieved from Twarc2}
}
\value{
Gives a quick summary of the tweet dataset `data`.
}
\description{
Quick Summary of Tweets
}
\examples{
quick_sum(twarc_data)
}
