% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{reduce_variable}
\alias{reduce_variable}
\title{Remove Unnecessary Variables}
\usage{
reduce_variable(data)
}
\arguments{
\item{data}{object that was retrieved from Twarc2}
}
\value{
This function reduces the number of columns from 75 to around 50. Use with caution. You may not want to remove all of them; but as far as I see, most people do not need these unneccessary variables.
}
\description{
Remove Unnecessary Variables
}
\examples{
quick_sum(twarc_data)
}
