% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{test_function}
\alias{test_function}
\title{Test Function}
\usage{
test_function(x)
}
\arguments{
\item{x}{Numeric input character vector}
}
\value{
A numeric vector the same length as `x`.
  `NA` strings have `NA` length.
}
\description{
Test Function
}
\examples{
test_function(6)
}
\seealso{
[nchar()]
}
